Refactor: move vuex modules files to their own folder
