+++ src/store/companies/companies-store.ts
@@ -1,8 +1,8 @@
 import { Module } from 'vuex';
 import getters from './company-getters';
 import state from './companies-state';
-import { BaseState } from './types';
-import { CompaniesState } from './companies-types';
+import { BaseState } from '@/types/types';
+import { CompaniesState } from '@/types/companies-types';
 
 export const companiesStore: Module<CompaniesState, BaseState> = {
   namespaced: true,
