+++ src/store/companies/company-getters.ts
@@ -1,11 +1,8 @@
 import { GetterTree } from 'vuex';
-import { CompaniesState } from './companies-types';
-import { BaseState } from './types';
+import { CompaniesState } from '@/types/companies-types';
+import { BaseState } from '@/types/types';
 
 const companiesGetters: GetterTree<CompaniesState, BaseState> = {
-  companies(state) {
-    return state.companies;
-  },
   companyTableHeaders(state) {
     return state.companyTableHeaders;
   },
